Fixes breakage caused by PyTorch 0.4.1
PyTorch 0.4.1 uses int64 as the default type for torch.arange if the
start/stop/step parameters are ints. This change caused tests to fail as
the torch.arange used to create the Ricker wavelet in tests became int64
instead of float.

+++ test/wavelets.py
@@ -5,7 +5,7 @@ import torch
 
 def ricker(freq, length, dt, peak_time):
     """Return a Ricker wavelet with the specified central frequency."""
-    t = torch.arange(length) * dt - peak_time
+    t = torch.arange(length, dtype=torch.float) * dt - peak_time
     y = (1 - 2 * math.pi**2 * freq**2 * t**2) \
         * torch.exp(-math.pi**2 * freq**2 * t**2)
     return y
